+++ src/gml-matrix/module.h
@@ -4,7 +4,7 @@
 #ifndef GML_MATRIX_MODULE_H
 #define GML_MATRIX_MODULE_H
 
-#include "../common.h"
+#include "common.h"
 #include <freetype2/ft2build.h>
 #include FT_FREETYPE_H
 
@@ -15,7 +15,7 @@
 #define CHAR_HEIGHT (GLYPH_HEIGHT + TEXT_PADDING)
 
 extern FT_Face face;
-extern int text_w, text_h;
+extern int32_t text_w, text_h;
 extern char* text_buffer; // Буфер размером text_w * text_h
 
 // Следующие функции вызываются в том порядке, в котором объявлены
@@ -26,7 +26,7 @@ void gml_read_config(config_t*);
 void gml_setup(void);
 
 /** Загружает и инициализирует ресурсы модуля после загрузки libdrm */
-void gml_setup_after_drm(uint32_t width, uint32_t height);
+void gml_setup_after_drm(uint16_t width, uint16_t height);
 
 
 /**
@@ -36,7 +36,7 @@ void gml_setup_after_drm(uint32_t width, uint32_t height);
  * @param height - высота фрейма.
  * @param frame - массив размером width * height, куда рендерится кадр.
  */
-void gml_draw(int tick, uint32_t width, uint32_t height, color_t* frame);
+void gml_draw(int tick, uint16_t width, uint16_t height, color_t* frame);
 
 
 /** Освобождает ресурсы модуля перед освобождением ресурсов libdrm */
