+++ src/gml-mint-logo/module.h
@@ -4,7 +4,7 @@
 #ifndef GML_MINT_LOGO_MODULE_H
 #define GML_MINT_LOGO_MODULE_H
 
-#include "../common.h"
+#include "common.h"
 #include <libpng/png.h>
 #include <freetype2/ft2build.h>
 #include FT_FREETYPE_H
@@ -27,7 +27,7 @@ void gml_read_config(config_t*);
 void gml_setup(void);
 
 /** Загружает и инициализирует ресурсы модуля после загрузки libdrm */
-void gml_setup_after_drm(uint32_t width, uint32_t height);
+void gml_setup_after_drm(uint16_t width, uint16_t height);
 
 
 /**
@@ -37,7 +37,7 @@ void gml_setup_after_drm(uint32_t width, uint32_t height);
  * @param height - высота фрейма.
  * @param frame - массив размером width * height, куда рендерится кадр.
  */
-void gml_draw(int tick, uint32_t width, uint32_t height, color_t* frame);
+void gml_draw(int tick, uint16_t width, uint16_t height, color_t* frame);
 
 
 /** Освобождает ресурсы модуля перед освобождением ресурсов libdrm */
