+++ src/gml-pixels/resources.c
@@ -3,7 +3,7 @@
  */
 
 #include "module.h"
-#include "../util/read_png.h"
+#include "util/read_png.h"
 #include <stdlib.h>
 #include <math.h>
 
@@ -28,9 +28,9 @@ void gml_setup(void) {
 	read_png(logo_path, &png_ptr, &info_ptr, &end_info);
 }
 
-void gml_setup_after_drm(uint32_t width, uint32_t height) {
-	uint32_t w = (width + PIXEL_SIZE - 1) / PIXEL_SIZE;
-	uint32_t h = (height + PIXEL_SIZE - 1) / PIXEL_SIZE;
+void gml_setup_after_drm(uint16_t width, uint16_t height) {
+	uint32_t w = ((uint32_t) width + PIXEL_SIZE - 1) / PIXEL_SIZE;
+	uint32_t h = ((uint32_t) height + PIXEL_SIZE - 1) / PIXEL_SIZE;
 	bitset2d_create(&buffer, w, h);
 
 	if (fabs(FPS - fps) > FPS_EPSILON) {
