+++ src/modules.h
@@ -5,10 +5,10 @@
 
 extern void (*read_config)(config_t*);
 extern void (*setup)(void);
-extern void (*setup_after_drm)(uint32_t, uint32_t);
+extern void (*setup_after_drm)(uint16_t, uint16_t);
 extern void (*cleanup_before_drm)(void);
 extern void (*cleanup)(void);
-extern void (*draw)(int, uint32_t, uint32_t, color_t*);
+extern void (*draw)(int, uint16_t, uint16_t, color_t*);
 
 void load_module(const char* filename);
